menu and menuBurger extracted to separate components
